figure-skating-bot/src/bot/handlers.py
# src/bot/handlers.py

from telegram import Update
from telegram.ext import CallbackContext

def start(update: Update, context: CallbackContext) -> None:
    update.message.reply_text('Привет! Я бот для анализа результатов по фигурному катанию.')

def help_command(update: Update, context: CallbackContext) -> None:
    update.message.reply_text('Используйте /start для начала.')

def analyze_results(update: Update, context: CallbackContext) -> None:
    # Логика для анализа результатов фигурного катания
    update.message.reply_text('Анализ результатов...')

def get_season_stats(update: Update, context: CallbackContext) -> None:
    # Логика для получения статистики за сезон
    update.message.reply_text('Получение статистики за сезон...')

def handle_pdf(update: Update, context: CallbackContext) -> None:
    # Логика для работы с PDF
    update.message.reply_text('Обработка PDF файлов...')