figure-skating-bot/requirements.txt
python
python-telegram-bot==13.7
pandas==1.3.3
numpy==1.21.2
PyPDF2==1.26.0
requests==2.26.0
sqlalchemy==1.4.22